services/auth-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
PyJWT==2.8.0
bcrypt==4.0.1
requests==2.31.0
gunicorn==21.2.0
cryptography==41.0.7
